+++ internal/generate/olm-catalog/csv_go_test.go
@@ -24,16 +24,13 @@ import (
 	"testing"
 
 	gen "github.com/operator-framework/operator-sdk/internal/generate/gen"
-	"github.com/operator-framework/operator-sdk/internal/scaffold"
 	"github.com/operator-framework/operator-sdk/internal/util/fileutil"
 	internalk8sutil "github.com/operator-framework/operator-sdk/internal/util/k8sutil"
 	"github.com/operator-framework/operator-sdk/internal/util/projutil"
-	"github.com/operator-framework/operator-sdk/pkg/k8sutil"
 
 	"github.com/blang/semver"
 	"github.com/ghodss/yaml"
 	"github.com/stretchr/testify/assert"
-	appsv1 "k8s.io/api/apps/v1"
 )
 
 const (
@@ -375,50 +372,5 @@ func TestUpdateVersion(t *testing.T) {
 	wantedReplaces := getCSVName(testProjectName, fromVersion)
 	if csv.Spec.Replaces != wantedReplaces {
 		t.Errorf("Wanted csv replaces %s, got %s", wantedReplaces, csv.Spec.Replaces)
-	}
-}
-
-func TestSetAndCheckOLMNamespaces(t *testing.T) {
-	cleanupFunc := chDirWithCleanup(t, testGoDataDir)
-	defer cleanupFunc()
-
-	depBytes, err := ioutil.ReadFile(filepath.Join(scaffold.DeployDir, "operator.yaml"))
-	if err != nil {
-		t.Fatalf("Failed to read Deployment bytes: %v", err)
-	}
-
-	// The test operator.yaml doesn't have "olm.targetNamespaces", so first
-	// check that depHasOLMNamespaces() returns false.
-	dep := appsv1.Deployment{}
-	if err := yaml.Unmarshal(depBytes, &dep); err != nil {
-		t.Fatalf("Failed to unmarshal Deployment bytes: %v", err)
-	}
-	if depHasOLMNamespaces(dep) {
-		t.Error("Expected depHasOLMNamespaces to return false, got true")
-	}
-
-	// Insert "olm.targetNamespaces" into WATCH_NAMESPACE and check that
-	// depHasOLMNamespaces() returns true.
-	setWatchNamespacesEnv(&dep)
-	if !depHasOLMNamespaces(dep) {
-		t.Error("Expected depHasOLMNamespaces to return true, got false")
-	}
-
-	// Overwrite WATCH_NAMESPACE and check that depHasOLMNamespaces() returns
-	// false.
-	overwriteContainerEnvVar(&dep, k8sutil.WatchNamespaceEnvVar, newEnvVar("FOO", "bar"))
-	if depHasOLMNamespaces(dep) {
-		t.Error("Expected depHasOLMNamespaces to return false, got true")
-	}
-
-	// Insert "olm.targetNamespaces" elsewhere in the deployment pod spec
-	// and check that depHasOLMNamespaces() returns true.
-	dep = appsv1.Deployment{}
-	if err := yaml.Unmarshal(depBytes, &dep); err != nil {
-		t.Fatalf("Failed to unmarshal Deployment bytes: %v", err)
-	}
-	dep.Spec.Template.ObjectMeta.Labels["namespace"] = olmTNMeta
-	if !depHasOLMNamespaces(dep) {
-		t.Error("Expected depHasOLMNamespaces to return true, got false")
 	}
 }