+++ internal/generate/olm-catalog/package_manifest_test.go
@@ -15,11 +15,6 @@
 package olmcatalog
 
 import (
-	"io/ioutil"
-	"log"
-	"os"
-	"path/filepath"
-	"reflect"
 	"testing"
 
 	"github.com/operator-framework/operator-registry/pkg/registry"
@@ -32,50 +27,33 @@ func TestGeneratePkgManifestToOutput(t *testing.T) {
 	cleanupFunc := chDirWithCleanup(t, testNonStandardLayoutDataDir)
 	defer cleanupFunc()
 
-	// Temporary output dir for generating catalog bundle
-	outputDir, err := ioutil.TempDir("", t.Name()+"-output-catalog")
-	if err != nil {
-		log.Fatal(err)
-	}
-	// Clean up output catalog dir
-	defer func() {
-		if err := os.RemoveAll(outputDir); err != nil && !os.IsNotExist(err) {
-			// Not a test failure since files in /tmp will eventually get deleted
-			t.Logf("Failed to remove tmp generated catalog directory (%s): %v", outputDir, err)
-		}
-	}()
-
 	cfg := gen.Config{
 		OperatorName: testProjectName,
-		OutputDir:    outputDir,
+		OutputDir:    "expected-catalog",
 	}
-
-	g := NewPackageManifest(cfg, csvVersion, "stable", true)
-	if err := g.Generate(); err != nil {
+	g := NewPackageManifest(cfg, csvVersion, "beta", false)
+	fileMap, err := g.(pkgGenerator).generate()
+	if err != nil {
 		t.Fatalf("Failed to execute package manifest generator: %v", err)
 	}
 
-	pkgManFileName := getPkgFileName(testProjectName)
-
-	// Read expected Package Manifest
-	expCatalogDir := filepath.Join("expected-catalog", OLMCatalogChildDir)
-	pkgManExpBytes, err := ioutil.ReadFile(filepath.Join(expCatalogDir, testProjectName, pkgManFileName))
-	if err != nil {
-		t.Fatalf("Failed to read expected package manifest file: %v", err)
+	if b, ok := fileMap[g.(pkgGenerator).fileName]; !ok {
+		t.Error("Failed to generate package manifest")
+	} else {
+		assert.Equal(t, packageManifestNonStandardExp, string(b))
 	}
-	pkgManExp := string(pkgManExpBytes)
-
-	// Read generated Package Manifest from OutputDir/olm-catalog
-	outputCatalogDir := filepath.Join(cfg.OutputDir, OLMCatalogChildDir)
-	pkgManOutputBytes, err := ioutil.ReadFile(filepath.Join(outputCatalogDir, testProjectName, pkgManFileName))
-	if err != nil {
-		t.Fatalf("Failed to read output package manifest file: %v", err)
-	}
-	pkgManOutput := string(pkgManOutputBytes)
-
-	assert.Equal(t, pkgManExp, pkgManOutput)
-
 }
+
+const packageManifestNonStandardExp = `channels:
+- currentCSV: memcached-operator.v0.0.1
+  name: alpha
+- currentCSV: memcached-operator.v0.0.3
+  name: beta
+- currentCSV: memcached-operator.v0.0.4
+  name: stable
+defaultChannel: stable
+packageName: memcached-operator
+`
 
 func TestGeneratePackageManifest(t *testing.T) {
 	cleanupFunc := chDirWithCleanup(t, testGoDataDir)
@@ -152,63 +130,6 @@ func TestValidatePackageManifest(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			if err := validatePackageManifest(tt.args.pkg); (err != nil) != tt.wantErr {
 				t.Errorf("Failed to check package manifest validate: error = %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
-func TestNewPackageManifest(t *testing.T) {
-	type args struct {
-		operatorName string
-		channelName  string
-		version      string
-	}
-	tests := []struct {
-		name string
-		args args
-		want registry.PackageManifest
-	}{
-		{
-			name: "Should return a valid registry.PackageManifest",
-			want: registry.PackageManifest{
-				PackageName: "memcached-operator",
-				Channels: []registry.PackageChannel{
-					registry.PackageChannel{
-						Name:           "stable",
-						CurrentCSVName: "memcached-operator.v0.0.3",
-					},
-				},
-				DefaultChannelName: "stable",
-			},
-			args: args{
-				operatorName: testProjectName,
-				channelName:  "stable",
-				version:      csvVersion,
-			},
-		},
-		{
-			name: "Should return a valid registry.PackageManifest with channel == alpha when it is not informed",
-			want: registry.PackageManifest{
-				PackageName: "memcached-operator",
-				Channels: []registry.PackageChannel{
-					registry.PackageChannel{
-						Name:           "alpha",
-						CurrentCSVName: "memcached-operator.v0.0.3",
-					},
-				},
-				DefaultChannelName: "alpha",
-			},
-			args: args{
-				operatorName: testProjectName,
-				version:      csvVersion,
-			},
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got := newPackageManifest(tt.args.operatorName, tt.args.channelName, tt.args.version)
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("NewPackageManifest() = %v, want %v", got, tt.want)
 			}
 		})
 	}